Completed
I developed a small project using .NET 5.0, Angular and SignalR technologies. My aim is to show real-time Covid19 case count using SignalR library. I developed two different projects to measure performance in the front-end part. You can see the Angular project on the left side of the screen and the .NET MVC project on the right. I coded the back-end part with .NET 5.0. When I send a request to the service, it randomly writes Covid19 cases for 5 cities into the database. As we write the data to the database, we can instantly see the number of cases in our graphic table.